contractcourt: add `Launch` method to htlc success resolver
This commit also refactors the `Resolve` method to give a clear flow on
how the output is being handled based on whether it's from the
local/remote commit and, whether it's using a pre/post anchor channel
type. Once this is done, it's easy to see what should be put in `Launch`
vs `Resolve`.