itest: shuffle test cases to even out blocks mined in tranches
This commit shuffles all the test cases before running them so tests
which require lots of blocks to be mined are less likely to be run in
the same tranch.

The other benefit is this approach provides a more efficient way to
figure which tests are broken since all the differnet backends are
running different tranches in their builds, we can identify more failed
tests in one push.