lntest+itest: remove the usage of `ht.AssertActiveHtlcs`
The method `AssertActiveHtlcs` is now removed due to it's easy to be
misused. To assert a given htlc, use `AssertOutgoingHTLCActive` and
`AssertIncomingHTLCActive` instead for ensure the HTLC exists in the
right direction. Although often the case `AssertNumActiveHtlcs` would be
enough has it implicitly checks the forwarding behavior for an
intermediate node by asserting there are always num_payment*2 HTLCs.