channeldb: increase db migration version
Migration 34 starts to disallow downgrading LND software if an
optional migration not known to the current LND version has already
been applied. Therefore we need to make sure when introducing this
change is mandatory so people cannot revert back because we cannot
rely on the meta data anymore for migration 34.